Move all end-to-end tests into one file.  Add end-to-end test for projection.
